Surcharge de la méthode equals() pour différencier les articles et les produits frais (remove(item))
