+++ src/model/Item.java
@@ -1,7 +1,5 @@
 package model;
 
-import java.util.Objects;
-
 public class Item {
     protected String name;
     protected long price;
@@ -48,14 +46,4 @@ public class Item {
         return this.name + ':' + ' ' + String.format("%.2f", this.price / 100.0);
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (o == this)
-            return true;
-        if (!(o instanceof Item))
-            return false;
-        Item item = (Item) o;
-        return Objects.equals(name, item.name) &&
-                price == item.price && weight == item.weight;
-    }
 }
